move tests to main_tests.cpp
