+++ tests/compile_main.sh
@@ -14,4 +14,4 @@ g++ -fpermissive -DF_CPU=8000000 -D__AVR_ATmega2560__ -I../speeduino -I./include
 g++ -D__AVR_ATmega2560__ -I../speeduino -I./include -c -o build/main.o main.cpp
 
 #link all that together into a working program
-g++ -o test build/Arduino.o build/FastCRC.o build/speeduino.o build/main.o build/PID.o lib/libgtest.a -lpthread
+g++ -o main build/Arduino.o build/FastCRC.o build/speeduino.o build/main.o build/PID.o lib/libgtest.a -lpthread
